+++ memio.h
@@ -1,13 +1,13 @@
 #ifndef _MEMIO_H
 #define _MEMIO_H
 
-#include "gftypes.h"
+#include "gfTypes.h"
 #include "tangerine.h"
 
-ulong mioInit( tangerineCtx_t *ptgctx );
+uint32_t mioInit( tangerineCtx_t *ptgctx );
 
-ulong fetchInstruction( ulong addr );
-ulong fetchData( ulong addr );
-ulong storeData( ulong addr, uchar mask, ulong data );
+uint32_t fetchInstruction( uint32_t addr );
+uint32_t fetchData( uint32_t addr );
+uint32_t storeData( uint32_t addr, uint8_t mask, uint32_t data );
 
 #endif
